requirements.txt
torch>=2.1
torchvision>=0.16
numpy>=1.24
pandas>=2.0
tqdm>=4.66
Pillow>=10.0
opencv-python>=4.8
pyyaml>=6.0
